Commit 1 to Frontend Branch
This commit contains UserInputSave.java as well as a user_tasks_save.txt file. UserInputSave.java is a GUI that has a text box and a submit button. Once the user inputs a task or time and submits, it will create a txt or save the text onto user_tasks_save.txt,